+++ bin/faucet/src/api/get_tokens.rs
@@ -109,11 +109,11 @@ pub enum MintRequestError {
     AssetAmountTooBig(AssetAmount, AssetAmount),
     #[error("requested amount {0} is not a valid asset amount")]
     InvalidAssetAmount(AssetAmountError),
-    #[error("POW error: {0}")]
+    #[error("PoW error")]
     PowError(#[from] PowError),
     #[error("API key {0} is invalid")]
     InvalidApiKey(String),
-    #[error("POW parameters are missing")]
+    #[error("PoW parameters are missing")]
     MissingPowParameters,
 }
 
@@ -128,6 +128,9 @@ pub enum GetTokenError {
 impl GetTokenError {
     fn status_code(&self) -> StatusCode {
         match self {
+            Self::InvalidRequest(MintRequestError::PowError(PowError::RateLimited)) => {
+                StatusCode::TOO_MANY_REQUESTS
+            },
             Self::InvalidRequest(_) | Self::MintError(_) => StatusCode::BAD_REQUEST,
             Self::FaucetOverloaded | Self::FaucetClosed => StatusCode::SERVICE_UNAVAILABLE,
             Self::FaucetReturnChannelClosed => StatusCode::INTERNAL_SERVER_ERROR,