+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added supply exceeded check ([#31](https://github.com/0xMiden/miden-faucet/pull/31)). 
 - Introduced `miden-faucet-client` crate ([#10](https://github.com/0xMiden/miden-faucet/pull/10)).
 - Replace amount options validation for maximum claimable amount ([#52](https://github.com/0xMiden/miden-faucet/pull/52)).
+- Use HTTP 429 status code for rate limited error ([#51](https://github.com/0xMiden/miden-faucet/pull/51)).
 
 ## 0.10.0 (2025-07-10)